UI checklist - Add 'Return to World Hub' button to pause menu etc
Introduced a new button in the pause menu for returning to the World Hub, updated button selection logic, and adjusted UI positions. Also refactored save manager settings to use integer volume values and updated main menu button event bindings for consistency.

+++ Plat_TCC/Assets/UI/Scripts/S_HelperFunctions.cs
@@ -11,7 +11,7 @@ public static class S_HelperFunctions
         return (((x - x_min) % (x_max - x_min)) + (x_max - x_min)) % (x_max - x_min) + x_min;
     }
 
-    public static float WrapVolume(float value, float step, float min, float max)
+    public static int WrapVolume(int value, int step, int min, int max)
     {
         value += step;
         if (value > max)
